move categories to lib
