+++ src/components/budget/provider.tsx
@@ -1,6 +1,6 @@
 'use client';
 import React, { createContext, useContext } from 'react';
-import { Category } from '@/data/categories';
+import { Category } from '@/lib/categories';
 import { CategoryId } from '@/types/shared';
 import { useBudgetForm } from '@/hooks/useBudgetForm';
 import { Project } from '@/__generated__/api/agora.schemas';
