+++ src/components/project-details/category-team.tsx
@@ -1,4 +1,4 @@
-import { type CategoryType, categoryNames } from '@/data/categories';
+import { type CategoryType, categoryNames } from '@/lib/categories';
 import { getBadgeClassName } from '@/lib/projectUtils';
 import { AvatarCarousel } from '../common/avatar-carousel';
 import { Badge } from '../ui/badge';
