+++ src/components/project-details/impact-statement.tsx
@@ -1,6 +1,6 @@
 import { ProjectImpactStatement } from '@/__generated__/api/agora.schemas';
 import { Heading } from '@/components/ui/headings';
-import { categoryNames } from '@/data/categories';
+import { categoryNames } from '@/lib/categories';
 import { Markdown } from '../markdown';
 
 export function ImpactStatement({
