+++ src/lib/categories.ts
